started the state machine node, still need to test the logic with simulator
